W4 completed

+++ internal/project/services/project_service.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"errors"
 	"time"
 
 	"github.com/google/uuid"
@@ -10,12 +11,14 @@ import (
 	"task-management/internal/project/repository"
 )
 
+var ErrForbidden = errors.New("you do not have permission to access this project")
+
 type ProjectService interface {
 	CreateProject(req dto.CreateProjectRequest, ownerID uuid.UUID) (*model.Project, error)
 	GetProjects() ([]*model.Project, error)
 	GetProjectByID(id uuid.UUID) (*model.Project, error)
-	UpdateProject(id uuid.UUID, req dto.UpdateProjectRequest) (*model.Project, error)
-	DeleteProject(id uuid.UUID) error
+	UpdateProject(id uuid.UUID, ownerID uuid.UUID, req dto.UpdateProjectRequest) (*model.Project, error)
+	DeleteProject(id uuid.UUID, ownerID uuid.UUID) error
 }
 
 type projectService struct {
@@ -55,10 +58,18 @@ func (s *projectService) GetProjectByID(id uuid.UUID) (*model.Project, error) {
 	return s.repo.GetByID(id)
 }
 
-func (s *projectService) UpdateProject(id uuid.UUID, req dto.UpdateProjectRequest) (*model.Project, error) {
+func (s *projectService) UpdateProject(
+	id uuid.UUID,
+	ownerID uuid.UUID,
+	req dto.UpdateProjectRequest,
+) (*model.Project, error) {
 	project, err := s.repo.GetByID(id)
 	if err != nil {
 		return nil, err
+	}
+
+	if project.OwnerID != ownerID {
+		return nil, ErrForbidden
 	}
 
 	project.Name = req.Name
@@ -72,6 +83,15 @@ func (s *projectService) UpdateProject(id uuid.UUID, req dto.UpdateProjectReques
 	return project, nil
 }
 
-func (s *projectService) DeleteProject(id uuid.UUID) error {
+func (s *projectService) DeleteProject(id uuid.UUID, ownerID uuid.UUID) error {
+	project, err := s.repo.GetByID(id)
+	if err != nil {
+		return err
+	}
+
+	if project.OwnerID != ownerID {
+		return ErrForbidden
+	}
+
 	return s.repo.Delete(id)
 }
